requirements.txt
requests==2.28.1
python-dotenv==0.20.0
Flask==2.3.3
Flask-CORS==4.0.0
markdown==3.4.4